initial refactor, including merging grasp objects. grasp sphere seems wrong
